Revert "Move service worker dataset context to a common space, to be used by notifications"
